pyproject.toml
[project]
name = "fyp"
version = "1.0.0"
description = "RAG-powered University FAQ WhatsApp Bot"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.20",
    "openai>=1.12.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "chromadb>=0.4.22",
    "tiktoken>=0.5.2",
    "flask>=3.0.0",
    "twilio>=8.12.0",
    "pydantic>=2.6.0",
    "langchain-chroma>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]
